% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHospitalName.R
\name{getHospitalName}
\alias{getHospitalName}
\title{Get hospital name from registry data}
\usage{
getHospitalName(reshID)
}
\arguments{
\item{reshID}{string defining the resh ID}
}
\value{
string of hospital name
}
\description{
Based on the hospital id (reshID) this function will return the name of
the corresponding hospital as provided in the registry data
}
\examples{
\dontrun{
getHospitalName("123456")
}
}
