% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRegDataTilsynsrapportMaaned.R
\name{getRegDataTilsynsrapportMaaned}
\alias{getRegDataTilsynsrapportMaaned}
\title{Provide dataframe of registry data}
\usage{
getRegDataTilsynsrapportMaaned(registryName, reshId)
}
\arguments{
\item{registryName}{String providing the current registryName}

\item{reshId}{String providing reshId for filtering in query}
}
\value{
regData data frame
}
\description{
Provides a dataframe containing data from a registry
}
